Update each.md for consistentcy
Fixes several instances in which the \@3each method was not highlighted/formatted as a method. While the explanation begins with using tic to indicate code, I think it is clear enough to change to the escaped # syntax once introduced in the headline "When to use \#each" as such. However, a broader discussion may be needed to determine a consistent writing style and stick to formatting style throughout the curriculum. (Sorry if I have missed it and it already exists.)